+++ lib/JCA_FNC_DigitalOut/JCA_FNC_DigitalOut.h
@@ -36,7 +36,7 @@ namespace JCA {
       uint16_t DelayCounter;
 
       // Intern
-      uint32_t LastMillis;
+      unsigned long LastMillis;
       uint16_t DelayMillis;
 
     public:
